Use marked as a dependency
