+++ index.html
@@ -28,7 +28,6 @@
         <script src='js/lib/bootstrap-tooltip.js'></script>
         <script src='js/lib/diff3.js'></script>
         <script src='js/lib/togeojson.js'></script>
-        <script src='js/lib/marked.js'></script>
 
         <script src='js/id/id.js'></script>
