+++ package.json
@@ -22,6 +22,7 @@
   ],
   "license": "ISC",
   "dependencies": {
+    "marked": "0.3.5",
     "osm-auth": "0.2.8",
     "rbush": "2.0.1",
     "sexagesimal": "0.5.0"
